requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.32
asyncpg==0.29.0
pydantic[email]==2.9.2
geoalchemy2==0.15.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
google-auth==2.34.0
httpx==0.27.2
alembic==1.13.2
psycopg2-binary==2.9.9
shapely==2.0.6
python-dotenv==1.0.1
